server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pgvector==0.2.5
openai==1.35.13
httpx==0.27.0
python-multipart==0.0.9
